+++ src/main/java/org/example/hatealcohol/controller/MainController.java
@@ -1,16 +1,16 @@
-package org.example.hatealcohol.Oauth2.controller;
+package org.example.hatealcohol.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
-public class MyController {
+public class MainController {
 
-    @GetMapping("/my")
+    @GetMapping("/")
     @ResponseBody
-    public String myAPI() {
+    public String mainAPI() {
 
-        return "my route";
+        return "main route";
     }
 }
