+++ src/main/java/org/example/hatealcohol/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package org.example.hatealcohol.Oauth2.dto;
+package org.example.hatealcohol.dto;
 
 public interface OAuth2Response {
