+++ src/main/java/org/example/hatealcohol/dto/UserDTO.java
@@ -1,4 +1,4 @@
-package org.example.hatealcohol.Oauth2.dto;
+package org.example.hatealcohol.dto;
 
 import lombok.Getter;
 import lombok.Setter;
